impletment expense creation and retreival

+++ internal/models/split.go
@@ -1,9 +1,9 @@
 package models
 
 type Split struct {
-	ID          int     `json:"id"`
-	Paid        float64 `json:"paid"`
-	Responsible float64 `json:"responsible"`
-	BillMateID  int     `json:"bill_mate_id"`
-	ExpenseID   int     `json:"expense_id"`
+	ID          int `json:"id"`
+	Paid        int `json:"paid"`
+	Responsible int `json:"responsible"`
+	BillMateID  int `json:"bill_mate_id"`
+	ExpenseID   int `json:"expense_id"`
 }
